requirements.txt
fastapi
requests
uvicorn
Jinja2
loguru
pydantic
starlette
sqlalchemy
python-jose
casbin
casbin_sqlalchemy_adapter
tencentcloud-sdk-python
passlib
bcrypt
pymysql